New Abstract class for sensors
