+++ CMakeLists.txt
@@ -82,6 +82,7 @@ INSTALL_SHARED_LIBRARY( ${PROJECT_NAME} DESTINATION ${PACKAGE_INSTALL_LIB_DIR} )
 
 ADD_SHARED_LIBRARY( MuonCVXDRealDigitiser   src/DetElemSlidingWindow.cc
                                             src/G4UniversalFluctuation.cc
+                                            src/AbstractSensor.cc
                                             src/HKBaseSensor.cc
                                             src/HitTemporalIndexes.cc
                                             src/MuonCVXDRealDigitiser.cc
